Revert "fix: move tests to lib directory"
This reverts commit 90caf0947ff9f696744475cf7db94522ba20af0c.
