+++ package.json
@@ -49,7 +49,7 @@
   "scripts": {
     "lint": "eslint '*/**/*.js'",
     "lint-fix": "eslint '*/**/*.js' --fix",
-    "test": "node ./node_modules/mocha/bin/_mocha 'dist/**/*.spec.js' --exit",
+    "test": "node ./node_modules/mocha/bin/_mocha './{,!(node_modules)/**/}*.spec.js' --exit",
     "test:ci": "yarn test",
     "start": "node dist/index.js",
     "version": "exit 0",
