+++ test/addNewMask.unit.spec.js
@@ -57,7 +57,7 @@ describe('addNewMask', () => {
             stubGetServerAddress.resolves(serverAddress);
             stubGot.post.resolves({ statusCode: 201 });
 
-            const { updateMasks, exitHandler } = proxyquire('../addNewMask', {
+            const { updateMasks, exitHandler } = proxyquire('../lib/addNewMask', {
                 './helpers': { getServerAddress: stubGetServerAddress },
             });
             process.listeners('exit').forEach((listener) => {
@@ -77,7 +77,7 @@ describe('addNewMask', () => {
     describe('negative', () => {
         it('should fail if the server address is not available', async () => {
             stubGetServerAddress.rejects('could not get server address');
-            const { updateMasks, exitHandler } = proxyquire('../addNewMask', {
+            const { updateMasks, exitHandler } = proxyquire('../lib/addNewMask', {
                 './helpers': {
                     getServerAddress: stubGetServerAddress,
                 },
@@ -94,7 +94,7 @@ describe('addNewMask', () => {
 
         it('should fail if the server address is not valid URL', async () => {
             stubGetServerAddress.resolves('foo');
-            const { updateMasks, exitHandler } = proxyquire('../addNewMask', {
+            const { updateMasks, exitHandler } = proxyquire('../lib/addNewMask', {
                 './helpers': {
                     getServerAddress: stubGetServerAddress,
                 },
@@ -116,7 +116,7 @@ describe('addNewMask', () => {
                 statusCode: 500,
                 body: 'Internal Server Error',
             });
-            const { updateMasks, exitHandler } = proxyquire('../addNewMask', {
+            const { updateMasks, exitHandler } = proxyquire('../lib/addNewMask', {
                 './helpers': { getServerAddress: stubGetServerAddress },
             });
             process.listeners('exit').forEach((listener) => {
@@ -132,7 +132,7 @@ describe('addNewMask', () => {
 
     describe('exitHandler', () => {
         it('should set exit code to 3 if the original exit code is 0 and variable was not masked', () => {
-            const { exitHandler } = proxyquire('../addNewMask', {});
+            const { exitHandler } = proxyquire('../lib/addNewMask', {});
             process.listeners('exit').forEach((listener) => {
                 if (listener === exitHandler) {
                     process.removeListener('exit', listener);
@@ -145,7 +145,7 @@ describe('addNewMask', () => {
         });
 
         it('should set exit code to 3 if the original exit code is 0 and variable was not masked', () => {
-            const { exitHandler } = proxyquire('../addNewMask', {});
+            const { exitHandler } = proxyquire('../lib/addNewMask', {});
             process.listeners('exit').forEach((listener) => {
                 if (listener === exitHandler) {
                     process.removeListener('exit', listener);
@@ -165,7 +165,7 @@ describe('addNewMask', () => {
             const serverAddress = 'https://xkcd.com/605/'
             stubGetServerAddress.resolves(serverAddress);
             stubGot.post.resolves({ statusCode: 201 });
-            const { updateMasks, exitHandler } = proxyquire('../addNewMask', {
+            const { updateMasks, exitHandler } = proxyquire('../lib/addNewMask', {
                 './helpers': { getServerAddress: stubGetServerAddress },
             });
             process.listeners('exit').forEach((listener) => {
