+++ test/isReady.unit.spec.js
@@ -1,6 +1,6 @@
 const chai = require('chai');
 const sinon = require('sinon');
-const { ContainerHandlingStatus } = require('../enums');
+const { ContainerHandlingStatus } = require('../lib/enums');
 const proxyquire = require('proxyquire');
 
 const expect = chai.expect;
@@ -21,7 +21,7 @@ describe('isReady script', () => {
         it('Should check exit with 0 code if container logger is ready', () => {
             const state = JSON.stringify({ status: 'ready', containers: {} })
             process.argv = [];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
@@ -31,7 +31,7 @@ describe('isReady script', () => {
         it('Should check exit with 1 code if container logger is not ready', () => {
             const state = JSON.stringify({ status: 'notReady', containers: {} })
             process.argv = [];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
@@ -43,7 +43,7 @@ describe('isReady script', () => {
         it('Should check exit with 0 code if container is ready', () => {
             const state = JSON.stringify({ status: 'ready', containers: { 'container-id': { status: ContainerHandlingStatus.LISTENING } } })
             process.argv = ['foo', 'bar', 'container-id'];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
@@ -53,7 +53,7 @@ describe('isReady script', () => {
         it('Should check exit with 0 code if container is waiting for start status', () => {
             const state = JSON.stringify({ status: 'ready', containers: { 'container-id': { status: ContainerHandlingStatus.WAITING_FOR_START } } })
             process.argv = ['foo', 'bar', 'container-id'];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
@@ -63,7 +63,7 @@ describe('isReady script', () => {
         it('Should check exit with 0 code if container is finished status', () => {
             const state = JSON.stringify({ status: 'ready', containers: { 'container-id': { status: ContainerHandlingStatus.FINISHED } } })
             process.argv = ['foo', 'bar', 'container-id'];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
@@ -73,7 +73,7 @@ describe('isReady script', () => {
         it('Should check exit with 1 code if container is not ready', () => {
             const state = JSON.stringify({ status: 'ready', containers: { 'container-id': { status: ContainerHandlingStatus.INITIALIZING } } })
             process.argv = ['foo', 'bar', 'container-id'];
-            proxyquire('../isReady.js', {
+            proxyquire('../lib/isReady.js', {
                 'fs': {
                     readFileSync: () => Buffer.from(state),
                 },
