+++ test/logger.unit.spec.js
@@ -5,9 +5,9 @@ const chai = require('chai');
 const sinon = require('sinon');
 const sinonChai = require('sinon-chai');
 const { EventEmitter } = require('events');
-const { ContainerStatus } = require('../enums');
-const { LoggerStrategy } = require('../enums');
-const { getPromiseWithResolvers } = require('../helpers');
+const { ContainerStatus } = require('../lib/enums');
+const { LoggerStrategy } = require('../lib/enums');
+const { getPromiseWithResolvers } = require('../lib/helpers');
 
 const expect = chai.expect;
 chai.use(sinonChai);
@@ -43,7 +43,7 @@ describe('Logger tests', () => {
     describe('constructor', () => {
 
         it('should define workflow log size limit default in case of non provided value', () => {
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -68,7 +68,7 @@ describe('Logger tests', () => {
         });
 
         it('should use passed workflow log size limit in case passed', () => {
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -111,7 +111,7 @@ describe('Logger tests', () => {
                     return Q.resolve(taskLogger);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                     'fastify': stubFastify,
                     './helpers': {
@@ -159,7 +159,7 @@ describe('Logger tests', () => {
                     return Q.resolve(taskLogger);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                     'fastify': stubFastify,
                     './helpers': {
@@ -205,7 +205,7 @@ describe('Logger tests', () => {
                     return Q.resolve(taskLogger);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                     'fastify': stubFastify,
                     './helpers': {
@@ -250,7 +250,7 @@ describe('Logger tests', () => {
                     return Q.resolve(taskLogger);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                     'fastify': stubFastify,
                     './helpers': {
@@ -300,7 +300,7 @@ describe('Logger tests', () => {
                     return Q.reject(new Error('my error'));
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                     'fastify': stubFastify,
                     './helpers': {
@@ -334,7 +334,7 @@ describe('Logger tests', () => {
                     callback(null, [{}, {}]);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'dockerode': function () {
                         return {
                             listContainers: listContainersSpy
@@ -362,7 +362,7 @@ describe('Logger tests', () => {
                     callback(null, []);
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'dockerode': function () {
                         return {
                             listContainers: listContainersSpy
@@ -394,7 +394,7 @@ describe('Logger tests', () => {
                     callback(new Error('getting containers error'));
                 });
 
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'dockerode': function () {
                         return {
                             listContainers: listContainersSpy
@@ -428,7 +428,7 @@ describe('Logger tests', () => {
         it('should call handleContainer in case of an create event', () => {
             const startSpy = sinon.spy();
             const onSpy = sinon.spy();
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'docker-events': function () {
                     return {
                         start: startSpy,
@@ -460,7 +460,7 @@ describe('Logger tests', () => {
                 callback();
             });
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fs': {
                     writeFile: writeFileSpy,
                     readFileSync: sinon.spy(() => {
@@ -485,7 +485,7 @@ describe('Logger tests', () => {
                 callback(new Error('write error'));
             });
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fs': {
                     writeFile: writeFileSpy,
                     readFileSync: sinon.spy(() => {
@@ -510,7 +510,7 @@ describe('Logger tests', () => {
     describe('validate', () => {
 
         it('should call process exit in case firebase authentication url was not provided', () => {
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -537,7 +537,7 @@ describe('Logger tests', () => {
             });
             process.exit = processExitSpy;
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -564,7 +564,7 @@ describe('Logger tests', () => {
             });
             process.exit = processExitSpy;
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -591,7 +591,7 @@ describe('Logger tests', () => {
             });
             process.exit = processExitSpy;
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'fastify': stubFastify,
                 './helpers': {
                     saveServerAddress: stubSaveServerAddress,
@@ -632,7 +632,7 @@ describe('Logger tests', () => {
                             emitter.start = startSpy;
                             return emitter;
                         });
-                        const Logger = proxyquire('../logger', {
+                        const Logger = proxyquire('../lib/logger', {
                             'cf-logs': {
                                 Logger: () => {
                                     return {
@@ -685,7 +685,7 @@ describe('Logger tests', () => {
                     it('should pass received step log size limit to ContainerLogger', async () => {
                         const infoSpy = sinon.spy();
                         const errorSpy = sinon.spy();
-                        const Logger = proxyquire('../logger', {
+                        const Logger = proxyquire('../lib/logger', {
                             'cf-logs': {
                                 Logger: () => {
                                     return {
@@ -737,7 +737,7 @@ describe('Logger tests', () => {
                     it('should pass undefined step log size limit to ContainerLogger in case of no label', async () => {
                         const infoSpy = sinon.spy();
                         const errorSpy = sinon.spy();
-                        const Logger = proxyquire('../logger', {
+                        const Logger = proxyquire('../lib/logger', {
                             'cf-logs': {
                                 Logger: () => {
                                     return {
@@ -785,7 +785,7 @@ describe('Logger tests', () => {
                     it('should update total log size when new log message event was sent', async () => {
                         const infoSpy = sinon.spy();
                         const errorSpy = sinon.spy();
-                        const Logger = proxyquire('../logger', {
+                        const Logger = proxyquire('../lib/logger', {
                             'cf-logs': {
                                 Logger: () => {
                                     return {
@@ -849,7 +849,7 @@ describe('Logger tests', () => {
                         });
                         const infoSpy = sinon.spy();
                         const errorSpy = sinon.spy();
-                        const Logger = proxyquire('../logger', {
+                        const Logger = proxyquire('../lib/logger', {
                             'cf-logs': {
                                 Logger: () => {
                                     return {
@@ -923,7 +923,7 @@ describe('Logger tests', () => {
             it('was previously handled', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -973,7 +973,7 @@ describe('Logger tests', () => {
             it('no loggerId', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1018,7 +1018,7 @@ describe('Logger tests', () => {
             it('no containerId', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1056,7 +1056,7 @@ describe('Logger tests', () => {
             it('no container status', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1094,7 +1094,7 @@ describe('Logger tests', () => {
             it('no step name', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1132,7 +1132,7 @@ describe('Logger tests', () => {
             it('no strategy provided', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1171,7 +1171,7 @@ describe('Logger tests', () => {
             it('provided strategy does not exist', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1211,7 +1211,7 @@ describe('Logger tests', () => {
             it('container status is create and strategy is logs', () => {
                 const infoSpy = sinon.spy();
                 const errorSpy = sinon.spy();
-                const Logger = proxyquire('../logger', {
+                const Logger = proxyquire('../lib/logger', {
                     'cf-logs': {
                         Logger: () => {
                             return {
@@ -1262,7 +1262,7 @@ describe('Logger tests', () => {
         it('should return true in case log limit exceeded', async () => {
             const infoSpy = sinon.spy();
             const errorSpy = sinon.spy();
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'cf-logs': {
                     Logger: () => {
                         return {
@@ -1315,7 +1315,7 @@ describe('Logger tests', () => {
         it('should return false in case log limit not exceeded', async () => {
             const infoSpy = sinon.spy();
             const errorSpy = sinon.spy();
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'cf-logs': {
                     Logger: () => {
                         return {
@@ -1376,7 +1376,7 @@ describe('Logger tests', () => {
         it('should return false in case log limit was not defined', async () => {
             const infoSpy = sinon.spy();
             const errorSpy = sinon.spy();
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'cf-logs': {
                     Logger: () => {
                         return {
@@ -1432,7 +1432,7 @@ describe('Logger tests', () => {
         it('should return true in case log limit exceeded for one container logger', async () => {
             const infoSpy = sinon.spy();
             const errorSpy = sinon.spy();
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'cf-logs': {
                     Logger: () => {
                         return {
@@ -1488,7 +1488,7 @@ describe('Logger tests', () => {
                 emitter.start = sinon.spy(() => Q.resolve());
                 return emitter;
             });
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 'cf-logs': {
                     Logger: () => {
                         return {
@@ -1560,7 +1560,7 @@ describe('Logger tests', () => {
             const containerLogger = new EventEmitter();
             containerLogger.start = () => Q.resolve();
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 '@codefresh-io/task-logger': { TaskLogger: () => Q.resolve(taskLogger) },
                 'docker-events': function () { return dockerEvents; },
                 './ContainerLogger': function () { return containerLogger; },
@@ -1654,7 +1654,7 @@ describe('Logger tests', () => {
             const containerLogger = new EventEmitter();
             containerLogger.start = () => Q.resolve();
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 '@codefresh-io/task-logger': { TaskLogger: () => Q.resolve(taskLogger) },
                 'docker-events': function () { return dockerEvents; },
                 './ContainerLogger': function () { return containerLogger; },
@@ -1696,7 +1696,7 @@ describe('Logger tests', () => {
             };
             const TaskLoggerFactory = sinon.spy(async () => taskLogger);
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                 'fastify': stubFastify,
                 './helpers': {
@@ -1739,7 +1739,7 @@ describe('Logger tests', () => {
             };
             const TaskLoggerFactory = sinon.spy(async () => taskLogger);
 
-            const Logger = proxyquire('../logger', {
+            const Logger = proxyquire('../lib/logger', {
                 '@codefresh-io/task-logger': { TaskLogger: TaskLoggerFactory },
                 'fastify': stubFastify,
                 './helpers': {
