+++ tsconfig.json
@@ -110,8 +110,8 @@
   },
   "exclude": [
     "node_modules",
-    // "**/*.unit*",
-    // "**/*.spec*",
+    "**/*.unit*",
+    "**/*.spec*",
     "coverage/**",
     "dist",
   ]
